Made a ElasticsearchBaseRepository in EchoCommon.
